Add doc/docx files choosing

+++ userbot/modules/documents.py
@@ -38,16 +38,24 @@ async def doc(e):
   if message.file.mime_type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document" or message.file.mime_type == "application/msword":
     file = message.document
     await e.edit("**Downloading...**")
-    file = await bot.download_file(file, "file.docx")
-    await e.edit("**Converting...**")
-    result = convertapi.convert('pdf', { 'File': 'file.docx' })
+    result = None
+    if  message.file.mime_type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document": #docx
+      file = await bot.download_file(file, "file.docx")
+      await e.edit("**Converting...**")
+      result = convertapi.convert('pdf', { 'File': 'file.docx' })
+      os.remove('file.docx')
+    if message.file.mime_type == "application/msword": #docx
+        file = await bot.download_file(file, "file.doc")
+        await e.edit("**Converting...**")
+        result = convertapi.convert('pdf', { 'File': 'file.doc' })
+        os.remove('file.doc')
     result.file.save('file.pdf')
     await e.edit("**Sending...**")
     await e.client.send_file(e.chat_id, f'file.pdf',reply_to=message)
-    os.remove('file.docx')
+    
     os.remove('file.pdf')
   else:
-    await e.edit("`Not a doc file. Aborting...`")
+    await e.edit("`Not a doc/docx file. Aborting...`")
     return
 @register(outgoing=True, pattern=r"^\.doc2img$")
 async def doc_png(e):
